+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -2,7 +2,7 @@ package wooteco.subway.dao;
 
 import java.util.List;
 
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 public interface LineDao {
