+++ src/main/java/wooteco/subway/service/StationService.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 import wooteco.subway.service.dto.ServiceDtoAssembler;
 import wooteco.subway.service.dto.station.StationResponseDto;
