+++ src/test/java/wooteco/subway/dao/FakeLineDao.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 
 public class FakeLineDao implements LineDao {
