+++ src/test/java/wooteco/subway/dao/FakeStationDao.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class FakeStationDao implements StationDao {
