+++ src/test/java/wooteco/subway/dao/jdbc/JdbcStationDaoTest.java
@@ -14,7 +14,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 
 import javax.sql.DataSource;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 @JdbcTest
 class JdbcStationDaoTest {
